requirements/project.txt

Mezzanine==0.11.10
sorl-thumbnail
south
django-pagination
MySQL-python
